requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# AWS Services
boto3==1.34.34

# Environment Management
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
Pillow==10.1.0
google-genai
requests==2.31.0
